Add shapes package, and shape and types modules
