+++ seacharts/shapes/__init__.py
@@ -1,0 +1,2 @@
+from .shape import Shape
+from .shapes import Area, Arrow, Circle, Line, Rectangle, Ship
