+++ seacharts/spatial/__init__.py
@@ -1,2 +1,1 @@
 from .hypsometry import Hydrography, Topography
-from .shapes import Area, Arrow, Circle, Line, Rectangle, Ship
